mesh/requirements.txt
numpy==1.22.4
Pillow==9.1.1
pycairo==1.21.0